Merge cellMarker signatures

Takes 2 cellMarkers signatures, merges them and recalculates optimal gene signatures.

Usage

mergeMarkers(
  mk1,
  mk2,
  remove_subclass = NULL,
  remove_group = NULL,
  transform = c("qq", "linear.qq", "scale", "none"),
  scale = 1,
  ...
)

Arguments

mk1

The reference 'cellMarkers' class object.

mk2

A 'cellMarkers' class object containing cell signatures to merge into mk1.

remove_subclass

Optional character vector of subclasses to remove when merging.

remove_group

Optional character vector of cell groups to remove when merging.

transform

Either "qq" which applies quantile_map() to mk2 to quantile transform it onto the same distribution as mk1, "linear.qq", which determines the quantile transformation and then applies a linear approximation of this, "scale" which simply scales the gene expression by the value scale, or "none" for no transformation.

scale

Numeric value determining the scaling factor for mk2 if transform is set to "scale".

...

Optional arguments and settings passed to updateMarkers().

Value

A list object of S3 class 'cellMarkers'. See cellMarkers() for details. If transform = "qq" then an additional element qqmerge is returned containing the quantile mapping function between the 2 datasets.

See also