Regression Deletion Diagnostics

Functions for computing regression diagnostics including standardised or Studentized residuals as well as Cook's distance.

Usage

# S3 method for class 'deconv'
rstudent(model, ...)

# S3 method for class 'deconv'
rstandard(model, ...)

# S3 method for class 'deconv'
cooks.distance(model, ...)

Arguments

model

'deconv' class object

...

retained for class compatibility

Value

Matrix of adjusted residuals or Cook's distance.

Details

Residuals are first adjusted for gene weights (if used). rstandard and rstudent give standardized and Studentized residuals respectively. Standardised residuals are calculated based on the hat matrix: $$H = X (X^T X)^{-1} X^T$$ Leverage \(h_{ii} = diag(H)\) is used to standardise the residuals: $$t_i = \cfrac{\hat{\varepsilon_i}}{\hat{\sigma} \sqrt{1 - h_{ii}}}$$ Studentized residuals are calculated based on excluding the \(i\) th case. Note this corresponds to refitting the regression, but without recomputing the non-negative compensation matrix. Cook's distance is calculated as: $$D_i = \cfrac{e_i^2}{ps^2} \left[\cfrac{h_{ii}}{(1 - h_{ii})^2} \right]$$ where \(p\) is the number of predictors (cell subclasses) and \(s^2\) is the mean squared error. In this model the intercept is not included.

See also